// libs/shared/modules/src/data/models/domain/customPolicy.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CustomPolicy {
    /**
     * The policy unique ID
     */
    readonly policy_id?: string;
    /**
     * The policy name
     */
    policy_name?: string;
    /**
     * The policy description
     */
    policy_description?: string;
    policy_type?: string;
}